// constants/enums/reserve_details_checkout.ts
export const enum ReserveDetailsCheckout_enum {
  NORMAL_WEEKDAYS = "weekdays",
  WEEK_ENDS = "weekends",
  PEAK_DAYS = "peaks",
  SPECIAL_DAYS = "specials",
  EXTRA_GUESTS = "extras",
  HOST_DISCOUNT = "host_discount",
  WEBSITE_DISCOUNT = "website_discount",
  RESERVE_PERIOD_DISCOUNT = "period_discount",
  COUPON_DISCOUNT = "coupon_discount",
}